feat: move access_field to expr.rs
