+++ core/engine/src/interpreter/expr.rs
@@ -1,7 +1,7 @@
 use log::debug;
-use roan_ast::{AccessKind, Assign, AssignOperator, BinOpKind, Binary, CallExpr, Expr, GetSpan};
+use roan_ast::{AccessKind, Assign, AssignOperator, BinOpKind, Binary, CallExpr, Expr, GetSpan, LiteralType};
 use roan_error::error::PulseError;
-use roan_error::error::PulseError::{UndefinedFunctionError, VariableNotFoundError};
+use roan_error::error::PulseError::{PropertyNotFoundError, UndefinedFunctionError, VariableNotFoundError};
 use crate::context::Context;
 use crate::module::{Module, StoredFunction};
 use crate::value::Value;
@@ -265,6 +265,49 @@ impl Module {
                 }
             },
             _ => todo!("missing left: {:?}", left),
+        }
+    }
+
+    /// Access a field of a value.
+    /// 
+    /// # Arguments
+    /// * `value` - The [Value] to access the field of.
+    /// * `expr` - The [Expr] representing the field to access.
+    /// * `ctx` - The context in which to access the field.
+    /// 
+    /// # Returns
+    /// The value of the field.
+    pub fn access_field(&mut self, value: Value, expr: &Expr, ctx: &Context) -> Result<Value> {
+        match expr {
+            Expr::Call(call) => {
+                let methods = value.builtin_methods();
+                if let Some(method) = methods.get(&call.callee) {
+                    let mut args = vec![value.clone()];
+                    for arg in call.args.iter() {
+                        self.interpret_expr(arg, ctx)?;
+                        args.push(self.vm.pop().expect("Expected value on stack"));
+                    }
+
+                    method.clone().call(args)
+                } else {
+                    Err(PropertyNotFoundError(call.callee.clone(), expr.span()).into())
+                }
+            }
+            Expr::Literal(lit) => {
+                if let LiteralType::String(s) = &lit.value {
+                    unimplemented!("There is not future that requires this code to be implemented now. This will be implemented with objects/structs.");
+                    // self.access_field(&Expr::Literal(lit.clone()))
+                } else {
+                    Err(PropertyNotFoundError("".to_string(), expr.span()).into())
+                }
+            }
+            _ => {
+                self.interpret_expr(expr, ctx)?;
+
+                let field = self.vm.pop().expect("Expected value on stack");
+
+                Ok(field)
+            }
         }
     }
 }
